style: Update styling of inputs
